(PC-8863) Refacto: put utilities in the right file
