+++ src/features/deeplinks/utils.ts
@@ -1,5 +1,9 @@
+import { omit } from 'search-params'
+
 import { DeepLinksToScreenConfiguration } from 'features/deeplinks/types'
 import { env } from 'libs/environment'
+
+import { DeeplinkEvent } from './types'
 
 export const DEEPLINK_DOMAIN = `https://${env.UNIVERSAL_LINK}/`
 export const FIREBASE_DYNAMIC_LINK_DOMAIN = `https://${env.FIREBASE_DYNAMIC_LINK}/`
@@ -32,6 +36,38 @@ export function generateLongFirebaseDynamicLink(
   customDynamicLinksParams = ''
 ) {
   return `${FIREBASE_DYNAMIC_LINK_DOMAIN}?link=${DEEPLINK_DOMAIN}${screen}?${universalLinksParams}&${getLongDynamicLinkURI()}${customDynamicLinksParams}`
+}
+
+export const isUniversalLink = (url: string) => url.startsWith(DEEPLINK_DOMAIN)
+export const isFirebaseDynamicLink = (url: string) => url.startsWith(FIREBASE_DYNAMIC_LINK_DOMAIN)
+export const isFirebaseLongDynamicLink = (url: string) =>
+  isFirebaseDynamicLink(url) && url.includes('?link=')
+
+/* For Firebase Dynamic Links with params (exemple /offer?id=234)
+ * we must use long dynamic links, and there are not recognized by dynamicLinks().onLink
+ * so we handle it manually
+ */
+export const extractUniversalLinkFromLongFirebaseDynamicLink = (event: DeeplinkEvent): string => {
+  const searchParams = event.url.replace(`${FIREBASE_DYNAMIC_LINK_DOMAIN}?`, '')
+  const paramsString = omit(searchParams, FIREBASE_DYNAMIC_LINK_PARAMS).querystring
+  return paramsString.replace(/^link=/, '')
+}
+
+export const resolveHandler = (handleDeeplinkUrl: (event: DeeplinkEvent) => void) => (
+  event: DeeplinkEvent
+) => {
+  if (isUniversalLink(event.url)) {
+    // Universal links: https://app.passculture-{env}.beta.gouv.fr/<routeName>
+    return handleDeeplinkUrl(event)
+  }
+
+  // Long Firebase Dynamic Links: https://passcultureapp{env}.page.link/?link=https://app.passculture-{env}.beta.gouv.fr/<routeName>?param=214906&apn=app.passculture.testing&isi=1557887412&ibi=app.passculture.test&efr=1
+  if (isFirebaseLongDynamicLink(event.url)) {
+    return handleDeeplinkUrl({ url: extractUniversalLinkFromLongFirebaseDynamicLink(event) })
+  }
+
+  // Short Firebase Dynamic Links: https://passcultureapp{env}.page.link/<routeName>
+  // => handled with dynamicLinks().onLink
 }
 
 /**
